backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
python-socketio==5.11.1
python-dotenv==1.0.1
Authlib==1.3.0
requests==2.31.0
